@@ -12,46 +12,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/cdk": "~7.3.7",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
+    "@angular/animations": "^8.2.3",
+    "@angular/cdk": "~8.1.0",
+    "@angular/common": "^8.2.3",
+    "@angular/compiler": "^8.2.3",
+    "@angular/core": "^8.2.3",
+    "@angular/forms": "^8.2.3",
+    "@angular/material": "^8.1.0",
+    "@angular/platform-browser": "^8.2.3",
+    "@angular/platform-browser-dynamic": "^8.2.3",
+    "@angular/router": "^8.2.3",
     "angular-in-memory-web-api": "^0.8.0",
     "core-js": "^2.6.9",
     "hammerjs": "^2.0.8",
     "jsonld": "^1.6.2",
     "rxjs": "^6.5.2",
-    "tslib": "^1.9.3",
-    "zone.js": "^0.8.29"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
-    "@types/jasmine": "~2.8.8",
+    "@angular-devkit/build-angular": "~0.803.0",
+    "@angular/cli": "^8.3.0",
+    "@angular/compiler-cli": "^8.2.3",
+    "@angular/language-service": "^8.2.3",
+    "@types/jasmine": "^3.4.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/jsonld": "^1.5.0",
     "@types/node": "~8.9.4",
     "angular-cli-ghpages": "^0.5.3",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "^3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.0",
+    "karma": "^4.0.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "^2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "~5.4.0",
-    "tar": ">=4.4.2",
     "ts-node": "^7.0.1",
     "tslint": "^5.16.0",
-    "typescript": "~3.2.4"
+    "typescript": "~3.5.3"
   }
 }
