@@ -1,9 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import {
-  MatListModule,
-} from '@angular/material';
+import { MatListModule } from '@angular/material/list';
 import { SimpleNaviComponent } from './simple-navi.component';
 
 @NgModule({
